Move a util class to test dir.
