+++ tsfile/src/main/java/org/apache/iotdb/tsfile/file/metadata/ChunkGroupMetaData.java
@@ -72,7 +72,7 @@ public class ChunkGroupMetaData {
    *
    * @param deviceID name of device
    * @param chunkMetaDataList all time series chunks in this chunk group. Can not be Null. notice:
-   * after constructing a ChunkGroupMetadata instance. Donot use list.add() to modify
+   * after constructing a ChunkGroupMetadata instance. Don't use list.add() to modify
    * `chunkMetaDataList`. Instead, use addTimeSeriesChunkMetaData() to make sure getSerializedSize()
    * is correct.
    * @param startOffsetOfChunkGroup the start Byte position in file of this chunk group.
